requirements.txt
absl-py==1.0.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
black==22.1.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.11
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
contextlib2==21.6.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
django-utils-six==2.0
entrypoints==0.4
execnet==1.9.0
executing==0.8.2
filelock==3.4.2
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
glob2==0.7
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.0.4
h5py==3.6.0
idna==2.8
importlib-metadata==4.10.1
iniconfig==1.1.1
ipykernel==6.8.0
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaggle==1.5.12
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mock==4.0.3
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.5
notebook==6.4.8
numba==0.55.1
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.5.5.62
opendatasets==0.1.20
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
path==16.3.0
path.py==12.5.0
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
pluggy==1.0.0
pooch==1.6.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.0
pytest-shutil==1.7.0
python-dateutil==2.8.2
python-dotenv==0.15.0
python-slugify==5.0.2
pytz==2021.3
pywin32==303
pywinpty==2.0.2
pyzmq==22.3.0
regex==2022.1.18
requests==2.22.0
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
scikit-learn==0.24.0
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
SoundFile==0.10.3.post1
spyder-kernels==2.2.1
sqlparse==0.4.2
stack-data==0.1.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
textblob==0.15.3
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tomli==2.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.25.11
virtualenv==20.13.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.2.0
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.7.0
